feat: todesign介绍页/直接搜索uid/用户空间&视频详情的异常处理/视频详情页的up主信息统一为UserCard()/小闪电的显示 del: 删除了一些冗余文件 fix: 从菜单页到缓存页没有动画

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/videoDetail/OwnerExt.kt
@@ -2,7 +2,7 @@ package cn.spacexc.wearbili.dataclass.videoDetail
 
 data class OwnerExt(
     val arc_count: String,
-    val assists: List<Int>,
+    val assists: List<Long>,
     val fans: Int,
     val live: Live,
     val official_verify: OfficialVerify,
